requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==9.0.0
groq==0.7.0
httpx==0.25.2
python-dotenv==1.0.0
